REV: Revert "REF: Move start and run commands into separate module"
This reverts commit b6595eab06a655a64068e1aaf26084180c35135d.
